Add IProductBuilder interface and ProductBuilder class
Introduced the `IProductBuilder` interface in the `Domain.Builders`
namespace to provide a fluent API for constructing `Product` objects.
The interface includes methods for setting various product properties
and a `Build` method to create the final object.

Added an empty `ProductBuilder` class, marked as `internal`, in the
same namespace. This class is intended to implement the `IProductBuilder`
interface in the future.